requirements.txt
pandas>=1.5.0
psycopg2-binary>=2.9.0
python-dateutil>=2.8.0
apache-airflow>=2.8.0
chardet>=5.0.0 